PedometerApp
This project is an Android application that provides a pedometer functionality to track steps and calculate calories burned. It stores user information in a Room Database, and uses SensorManager and SensorEventListener to get data from the step counting sensor. The application calculates and displays the calories burned with each step. It utilizes modern Android technologies like Kotlin, LifecycleScope, and Coroutine.